[BUGFIX] IndexOfAny was calling the first_index_of wrongly.
I misinterpreted the meaning the function and the call is
assuming that count argument is related with the str argument
while it was related with the chars argument.

While this is not a performatic way to achieve the result,
at least it's correct (I think), so I let a COWTODO
to check it later.